// src/api/__generated__/story/model/listPublicDramasSort.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */

export type ListPublicDramasSort =
  (typeof ListPublicDramasSort)[keyof typeof ListPublicDramasSort];

export const ListPublicDramasSort = {
  hot: 'hot',
  latest: 'latest',
  favorite: 'favorite',
  completed_view: 'completed_view',
} as const;